Refactoring Stochastic classes to their own files.
