+++ installCheck.py
